Views added and CRUD implemented

+++ src/main/java/com/schedule/mvc/dao/IGenericDao.java
@@ -12,9 +12,9 @@ public interface IGenericDao<E> {
 
     public List<E> selectAll(String sortField, String sort);
 
-    public E getById(int id);
+    public E getById(Long id);
 
-    public void remove(int id);
+    public void remove(Long id);
 
     public E findOne(String paramName, Object paramValue);
