+++ src/main/java/com/schedule/mvc/services/IRouteService.java
@@ -14,9 +14,9 @@ public interface IRouteService {
 
     public List<Route> selectAll(String sortField, String sort);
 
-    public Route getById(int id);
+    public Route getById(Long id);
 
-    public void remove(int id);
+    public void remove(Long id);
 
     public Route findOne(String paramName, Object paramValue);
